+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New and Changed Functionality
 
 - Added option to select how data should be sorted when displayed on the plot
+- Added dropdown boxes to select how many rows and columns are desired
 
 ### Miscellaneous Changes